Use automatic scala version in launcher properties
This is intended to fix #534 where if we try to launch
an older version of activator we give it the too-new
Scala; it would also in theory fix a future problem
where if we try to launch a newer-than-us activator on
a newer scala things wouldn't work.

My only misgiving is that I guess there's some reason we didn't
do this before (some reason we hardcoded the scala version before)
but I guess we will find out.